Add touch+mouse controls, rewrite README, and upgrade fs precision.

+++ fs.js
@@ -1,5 +1,5 @@
 const fs = 
-`precision mediump float;
+`precision highp float;
 
 uniform vec2 iResolution;
 uniform mat3 cam;
